requirements.txt
diffusers>=0.5.1
numpy==1.23.4
torch
torchvision
transformers>=4.21.0
Pillow==9.2.0
tqdm==4.64.1
ftfy==6.1.1
bitsandbytes
accelerate==0.13.1
scipy==1.9.3
hivemind
pytorch_lightning
OmegaConf